Разнесение логики и ассетов и настроек игры по отдельным папкам
