+++ packages/client/src/pages/Game/GameLogic/assets/background_0.png
