+++ packages/client/src/pages/Game/GameLogic/assets/background_2.png
