+++ packages/client/src/pages/Game/GameLogic/assets/warrior/Attack2.png
