+++ packages/client/src/pages/Game/GameLogic/assets/warrior/Death.png
