+++ packages/client/src/pages/Game/GameLogic/assets/warrior/Idle.png
