+++ packages/client/src/pages/Game/GameLogic/assets/warrior/IdleLeft.png
