+++ packages/client/src/pages/Game/GameLogic/assets/warrior/Jump.png
