+++ packages/client/src/pages/Game/GameLogic/assets/warrior/JumpLeft.png
