+++ packages/client/src/pages/Game/GameLogic/assets/warrior/Run.png
