+++ packages/client/src/pages/Game/GameLogic/assets/warrior/Take Hit - white silhouette.png
