+++ packages/client/src/pages/Game/GameLogic/classes/Sprite.ts
@@ -1,4 +1,4 @@
-import { ICoordinates } from '../models';
+import { ICoordinates } from '../../models';
 
 interface ISpiteConstructor {
     position: ICoordinates;
