+++ packages/client/src/pages/Game/GameLogic/configs/keys.ts
@@ -1,0 +1,11 @@
+export const GAME_KEYS = {
+    d: {
+        pressed: false,
+    },
+    a: {
+        pressed: false,
+    },
+    w: {
+        pressed: false,
+    },
+};
