+++ packages/client/src/pages/Game/GameLogic/configs/playerConfig.ts
@@ -1,0 +1,8 @@
+export { default as WarriorIdle } from '../assets/warrior/Idle.png';
+export { default as WarriorIdleLeft } from '../assets/warrior/IdleLeft.png';
+export { default as WarriorRun } from '../assets/warrior/Run.png';
+export { default as WarriorJump } from '../assets/warrior/Jump.png';
+export { default as WarriorJumpLeft } from '../assets/warrior/JumpLeft.png';
+export { default as WarriorFall } from '../assets/warrior/Fall.png';
+export { default as WarriorFallLeft } from '../assets/warrior/FallLeft.png';
+export { default as WarriorRunLeft } from '../assets/warrior/RunLeft.png';
