+++ packages/client/src/pages/Game/GameLogic/index.module.scss
@@ -1,0 +1,52 @@
+$game-start-bg-color: rgba(0, 0, 0, 0.3);
+$game-start-wrap-gradient-40: rgba(194, 194, 194, 0.596);
+$game-start-wrap-border-color: rgba(0, 0, 0, 0.404);
+$game-start-menu-shadow-color: rgba(21, 21, 21, 0.623);
+
+.game-start {
+    position: absolute;
+    top: 0;
+    left: 0;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100vh;
+    width: 100vw;
+    background-color: $game-start-bg-color;
+    color: var(--black);
+    text-align: center;
+    backdrop-filter: blur(1.5px);
+    -webkit-backdrop-filter: blur(1.5px);
+
+    .wrap {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        background: radial-gradient(var(--gray-light) 50%, var(--blue) 140%);
+        border: 1px solid $game-start-wrap-border-color;
+        border-radius: 32px;
+        box-shadow: 10px 10px 0px $game-start-menu-shadow-color;
+
+        button {
+            width: max-content;
+            background-color: var(--main-page-btn-bg-color);
+            color: var(--white);
+            padding: 0.8rem 2rem;
+            border: none;
+            border-radius: 0;
+            box-shadow: 6px 6px 0px $game-start-menu-shadow-color;
+
+            &:hover {
+                background-color: var(--orange-hover);
+            }
+
+            &:active {
+                background-color: var(--orange-hover);
+                transform: translate(6px, 6px);
+                box-shadow: none;
+            }
+        }
+    }
+}
