--- packages/client/src/pages/Game/consts.ts
@@ -1,2 +1,0 @@
-export const CANVAS_WIDTH = 1280;
-export const CANVAS_HEIGHT = 900;
